refactor(cli): collapse build into run; rename ad-hoc command verb to exec
The CLI now exposes one production verb for actions, plus a substrate
escape hatch:

- fabrik run //pkg:name: execute the action(s) that produce the named
  target. Uniform across target kinds; for a rust_binary it runs rustc.
- fabrik exec -- argv: cache and execute a literal command without
  touching the target graph (substrate-level escape hatch). Renamed
  from the previous fabrik run.

The old fabrik build verb is gone. At the action layer there is no
distinction between "build" and "run": producing a target's outputs is
running its declared action(s); whether the result also gets executed
afterward is encoded in the target type, not the CLI verb. Keeps the
surface small for agents (one verb to learn) and pushes behavioural
variation into the typed graph.

Touched:
- cli.rs: Cmd::Build -> Cmd::Run { label }; Cmd::Run -> Cmd::Exec.
- commands/build.rs renamed to commands/run.rs (run target);
  commands/run.rs renamed to commands/exec.rs (substrate).
- spec/build_spec.sh -> spec/run_spec.sh; spec/run_spec.sh ->
  spec/exec_spec.sh; cli_surface_spec.sh and cache_stats_spec.sh
  updated to reflect the new verb names.
- README quick-taste section now leads with fabrik run //hello:hello
  and shows fabrik exec as the substrate path.
- docs/design.md gains section 6.4 codifying the one-verb design.

Coverage: 39/39 shellspec, 10/10 frontend unit tests, clippy and fmt
clean.
