+++ crates/fabrik-cli/src/commands/mod.rs
@@ -2,7 +2,7 @@
 //! dispatcher in [`crate::main`] routes parsed [`crate::cli::Cmd`] into
 //! these.
 
-pub mod build;
 pub mod cache;
+pub mod exec;
 pub mod run;
 pub mod targets;
